Fix and update rating calculation

+++ Models/Review.cs
@@ -21,7 +21,7 @@ namespace VNPT_Review.Models
 
         [Required]
         [Display(Name = "Mức đánh giá")]
-        public decimal Rating { get; set; }
+        public float Rating { get; set; }
 
         [Required]
         [Display(Name = "Nội dung đánh giá")]
